Changes were done to give mobility to fovea
Was built the second version to testfovea, cause was necessary to
test how modify the fovea position. This was done without problems,
but it's necessary to change the show function in fovea, cause it
doesn't working like should be, in other words, the levels are being
showed and in this case it's necessary to show the image foveated
with features being extracted.