Match function and graph class
In this commit was created a match function inside the fovea class with
the goal of define which and how many matches were made between model
and scene analised. Furthermore, there are two possibilities to make the
match in this function and these can be analised using the Graph class,
ploting graphics 2D/3D, showed in testdetectGraph.cpp file. I recommend
reading the menu function ( inside testdetectGraph.cpp file ) before
compiling and running this test.