3D Graphic representing all levels and Statistic class
The fovea was positioned in all positions of image and the values of
proportion of inliers were save and plotted by levels. This allowed
to visualize the distribution by levels, but to do it was necessary
to make a filter, because there were high proportion of inliers with
few matches. In addition, the statistical class was implemented to
adjust weights of inlier proportions per level to a single value and
functions that assist in the convergence of an estimated position to
the target sought.