Testing the level weighting function
The weight function of the levels contributes perfectly to constructing
the fovea weight value, but there is a time performance problem that has
been detected in the findHomography function. The performance of multiple
fovea falls dramatically after the use of this function and it was also
noticed that the ORB feature parameters also contribute to the fovea size
construction. That is, it is necessary to decrease the parameters "edgeThreshold"
and "patchSize" to be able to decrease the size of the fovea.