Adding foveatedFeatures Codes into the calcLayerDetAndTrace Function
This commit used the fovea structure to update all parameters of the
calcLayerDetAndTrace function from the foveatedHessianDetector file.
However, in this version there is a problem in the generation of
keypoints through the method proposed by Rafael. We believe the error
is in the "K fovea" parameter, which is passed as an argument in the
SURFFindInvoker structure.