Removal of methods that do not choose the shape of the fovea
The shape of the fovea should not be overlooked, because it can
represent several optimizations in the future of computer graphics
( the detail is that this is already happening ) and computer vision
models. With that in mind, it is up to the system user to choose the
appropriate shape for their context.