Datasets foveated
An algorithm with the fovea class was built, where a dataset directory
is passed as argument and it transforms the images into centralized
foveae. An interesting strategy to be implemented is to use the region
defined in the training file as a position for the fovea that must be
fed with inputs for training the neural network.