Adding Error Graph Between Configuration and Methods
The calibration.py file is used to find intrinsics parameters that
will be used to estimate the position of target on the image. Using
the parameters generates by calibration.py we create GraphicError.dat
file contains all data necessary to generate the graphic error vs image.
It's important to remember that was used Toy dataset available at
link: http://cvlab.hanyang.ac.kr/tracker_benchmark/index.html